+++ inc/woocommerce/hooks.php
@@ -17,6 +17,7 @@ add_filter( 'woocommerce_enqueue_styles', 	'__return_empty_array' );
  * @see  storefront_before_content()
  * @see  storefront_after_content()
  * @see  woocommerce_breadcrumb()
+ * @see  storefront_shop_messages()
  */
 remove_action( 'woocommerce_before_main_content', 	'woocommerce_breadcrumb', 					20, 0 );
 remove_action( 'woocommerce_before_main_content', 	'woocommerce_output_content_wrapper', 		10 );
@@ -27,6 +28,7 @@ remove_action( 'woocommerce_before_shop_loop', 		'woocommerce_result_count', 			
 remove_action( 'woocommerce_before_shop_loop', 		'woocommerce_catalog_ordering', 			30 );
 add_action( 'woocommerce_before_main_content', 		'storefront_before_content', 				10 );
 add_action( 'woocommerce_after_main_content', 		'storefront_after_content', 				10 );
+add_action( 'storefront_content_top', 				'storefront_shop_messages', 				1 );
 add_action( 'storefront_content_top', 				'woocommerce_breadcrumb', 					10 );
 
 add_action( 'woocommerce_after_shop_loop',			'storefront_sorting_wrapper',				9 );
@@ -41,8 +43,6 @@ add_action( 'woocommerce_before_shop_loop', 		'woocommerce_result_count', 				20
 add_action( 'woocommerce_before_shop_loop', 		'woocommerce_pagination', 					30 );
 add_action( 'woocommerce_before_shop_loop',			'storefront_sorting_wrapper_close',			31 );
 
-
-
 /**
  * Products
  * @see  storefront_upsell_display()
@@ -51,7 +51,6 @@ remove_action( 'woocommerce_after_single_product_summary', 	'woocommerce_upsell_
 add_action( 'woocommerce_after_single_product_summary', 	'storefront_upsell_display', 				15 );
 remove_action( 'woocommerce_before_shop_loop_item_title', 	'woocommerce_show_product_loop_sale_flash', 10 );
 add_action( 'woocommerce_after_shop_loop_item_title', 		'woocommerce_show_product_loop_sale_flash', 6 );
-
 
 /**
  * Header