+++ changelog.md
@@ -5,6 +5,7 @@
 * **Fix** - Third level dropdowns (and beyond) are now revealed in the correct situation.
 * **Fix** - Header margin on homepage template when WooCommerce isn't activated.
 * **Fix** - Dropdowns on touch devices.
+* **Fix** - Pagination hidden correctly when only displaying categories / subcategories.
 * **Tweak** - Output WooCommerce messages on all appropriate pages.
 
 ### *2015.04.08* - 1.4.3